requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.32
boto3==1.26.0